chore: IntentUtils
